chore(organisation): reorganise repo structure - move reusable code to helpers folder - separate logic in controllers to workers folder [Delivers #161816772]
