+++ .eslintrc.json
@@ -20,9 +20,9 @@
         "requireReturn": true,
         "requireReturnType": true,
         "requireParamDescription": false,
-        "requireReturnDescription": true,
-        "max-line-length": [true, 100]
+        "requireReturnDescription": true
         }],
+        "max-line-length": [true, 100],
         "require-jsdoc": ["error", {
             "require": {
                 "FunctionDeclaration": true,
