--- server/src/helpers/helpers.js
@@ -1,5 +1,0 @@
-import addTags from './addTags';
-
-export default {
-  addTags,
-};
