+++ server/src/helpers/index.js
@@ -1,0 +1,15 @@
+import addTags from '../workers/TagWorker';
+import issueToken from './issueToken';
+import multifile from './multifile';
+import resetPasswordTemplate from './resetPasswordTemplate';
+import sendMail from './sendMail';
+import slugDecoder from './slugDecoder';
+
+export default {
+  addTags,
+  issueToken,
+  multifile,
+  resetPasswordTemplate,
+  sendMail,
+  slugDecoder,
+};
