+++ server/src/middlewares/validators.js
@@ -1,0 +1,21 @@
+import postArticle from './articleValidations/postArticle';
+import search from './articleValidations/search';
+import tags from './articleValidations/tags';
+import createProfile from './userValidations/createProfile';
+import forgotPassword from './userValidations/forgotPassword';
+import login from './userValidations/login';
+import resetPassword from './userValidations/resetPassword';
+import signup from './userValidations/signup';
+import token from './userValidations/token';
+
+export default {
+  postArticle,
+  search,
+  tags,
+  createProfile,
+  forgotPassword,
+  login,
+  resetPassword,
+  signup,
+  token,
+};
