+++ server/src/routes/api/search.js
@@ -1,9 +1,9 @@
 import express from 'express';
 import SearchController from '../../controllers/SearchController';
-import searchValidator from '../../helpers/validationHelpers/searchValidation';
+import Validators from '../../middlewares/validators';
 
 const router = express.Router();
 
-router.post('/search', searchValidator, SearchController.search);
+router.post('/search', Validators.search, SearchController.search);
 
 export default router;
