+++ server/src/test/index.js
@@ -1,9 +1,9 @@
-import userTest from './userTest';
-import articleTest from './articleTest';
-import searchTest from './searchTest';
-import tagsTest from './tagsTest';
-import paginationTest from './paginationTest';
-import reactionTest from './reactionTest';
+import userTest from './integration tests/userTest';
+import articleTest from './integration tests/articleTest';
+import searchTest from './integration tests/searchTest';
+import tagsTest from './integration tests/tagsTest';
+import paginationTest from './integration tests/paginationTest';
+import reactionTest from './integration tests/reactionTest';
 
 export default {
   userTest,
