+++ server/src/test/integration tests/tagsTest.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../../../index';
+import app from '../../../../index';
 
 const { expect } = chai;
 chai.use(chaiHttp);
